refactor: move runtime directory to top-level
