+++ packages/vite/src/node/__tests__/tsconfig.json
@@ -1,6 +1,5 @@
 {
-  "extends": "./tsconfig.base.json",
-  "include": ["./rollup.config.ts", "./rollup.dts.config.ts"],
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
     "esModuleInterop": true,
     "declaration": false,
