+++ packages/vite/src/node/index.ts
@@ -128,10 +128,10 @@ export type {
   HMRBroadcasterClient,
 } from './server/hmr'
 
-export type { FetchFunction } from './ssr/runtime/index'
-export { createViteRuntime } from './ssr/runtime/node/mainThreadRuntime'
-export type { MainThreadRuntimeOptions } from './ssr/runtime/node/mainThreadRuntime'
-export { ServerHMRConnector } from './ssr/runtime/node/serverHmrConnector'
+export type { FetchFunction } from '../runtime/index'
+export { createViteRuntime } from './ssr/runtime/mainThreadRuntime'
+export type { MainThreadRuntimeOptions } from './ssr/runtime/mainThreadRuntime'
+export { ServerHMRConnector } from './ssr/runtime/serverHmrConnector'
 
 export type { BindCLIShortcutsOptions, CLIShortcut } from './shortcuts'
