+++ packages/vite/src/node/server/index.ts
@@ -48,7 +48,7 @@ import { printServerUrls } from '../logger'
 import { createNoopWatcher, resolveChokidarOptions } from '../watch'
 import { initPublicFiles } from '../publicDir'
 import { getEnvFilesForMode } from '../env'
-import type { FetchResult } from '../ssr/runtime/types'
+import type { FetchResult } from '../../runtime/types'
 import { ssrFetchModule } from '../ssr/ssrFetchModule'
 import type { PluginContainer } from './pluginContainer'
 import { ERR_CLOSED_SERVER, createPluginContainer } from './pluginContainer'
