+++ packages/vite/src/runtime/__tests__/fixtures/assets.js
