+++ packages/vite/src/runtime/__tests__/fixtures/basic.js
