+++ packages/vite/src/runtime/__tests__/fixtures/c.ts
@@ -1,4 +1,4 @@
-/* eslint-disable no-console */
+ 
 
 export { a as c } from './a'
