+++ packages/vite/src/runtime/__tests__/fixtures/circular/circular-a.js
