+++ packages/vite/src/runtime/__tests__/fixtures/circular/circular-b.js
