+++ packages/vite/src/runtime/__tests__/fixtures/cjs-external/index.cjs
