+++ packages/vite/src/runtime/__tests__/fixtures/d.ts
@@ -1,4 +1,4 @@
-/* eslint-disable no-console */
+ 
 
 export { c as d } from './c'
