+++ packages/vite/src/runtime/__tests__/fixtures/dynamic-import.js
