+++ packages/vite/src/runtime/__tests__/fixtures/esm-external-existing.js
