+++ packages/vite/src/runtime/__tests__/fixtures/esm-external-non-existing.js
