+++ packages/vite/src/runtime/__tests__/fixtures/esm-external/index.mjs
