+++ packages/vite/src/runtime/__tests__/fixtures/has-error.js
