+++ packages/vite/src/runtime/__tests__/fixtures/hmr.js
