+++ packages/vite/src/runtime/__tests__/fixtures/installed.js
