+++ packages/vite/src/runtime/__tests__/fixtures/native.js
