+++ packages/vite/src/runtime/__tests__/fixtures/throws-error-method.ts
