+++ packages/vite/src/runtime/__tests__/fixtures/virtual.js
