+++ packages/vite/src/runtime/__tests__/package.json
