+++ packages/vite/src/runtime/__tests__/server-runtime.spec.ts
