+++ packages/vite/src/runtime/constants.ts
