+++ packages/vite/src/runtime/sourcemap/interceptor.ts
