+++ packages/vite/src/runtime/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "include": ["./", "../node", "../dep-types", "../types"],
+  "exclude": ["**/__tests__"],
+  "compilerOptions": {
+    "lib": ["ESNext", "DOM"],
+    "stripInternal": true
+  }
+}
