+++ packages/vite/src/runtime/utils.ts
